refactor: 서버 선택 관련 로직을 `ServerSelector` 컴포넌트로 분리 및 적용
